backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.121.4
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
supabase==2.13.2

# AI/ML
openai==1.59.5
anthropic==0.40.0
numpy==2.2.1
scikit-learn==1.6.1
pandas==2.2.3

# PDF Generation
reportlab==4.2.5
matplotlib==3.10.0
seaborn==0.13.2

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
email-validator==2.2.0
httpx==0.28.1

# Development
pytest==8.3.4
pytest-asyncio==0.25.2
black==24.11.0
ruff==0.8.6
mypy==1.14.1

# Monitoring
prometheus-client==0.21.1
sentry-sdk[fastapi]==2.20.1